+++ lib/hobbes/util/region.C
@@ -8,14 +8,14 @@ namespace hobbes {
 
 void dbglog(const std::string&);
 
-region::region(unsigned int minPageSize, unsigned int initialFreePages, unsigned int maxPageSize) :
+region::region(size_t minPageSize, size_t initialFreePages, size_t maxPageSize) :
   minPageSize(minPageSize), maxPageSize(maxPageSize), lastAllocPageSize(minPageSize),
   usedp(0), freep(0),
   abortOnOOM(false), maxTotalAllocation(0), totalAllocation(0)
 {
   this->usedp = newpage(0, minPageSize);
 
-  for (unsigned int i = 0; i < initialFreePages; ++i) {
+  for (size_t i = 0; i < initialFreePages; ++i) {
     this->freep = newpage(this->freep, minPageSize);
   }
 }
@@ -25,8 +25,8 @@ region::~region() {
   freepage(this->usedp);
 }
 
-void* region::malloc(unsigned int sz) {
-  unsigned int nr = this->usedp->read + sz;
+void* region::malloc(size_t sz) {
+  size_t nr = this->usedp->read + sz;
   if (nr <= this->usedp->size) {
     void* result = ((unsigned char*)this->usedp->base) + this->usedp->read;
     this->usedp->read = nr;
@@ -141,7 +141,7 @@ void region::abortAtMemCeiling(size_t maxsz) {
   this->maxTotalAllocation = maxsz;
 }
 
-mempage* region::newpage(mempage* succ, unsigned int sz) {
+mempage* region::newpage(mempage* succ, size_t sz) {
   size_t psz = 0;
   if (this->lastAllocPageSize < this->maxPageSize) {
     psz = std::max(sz, this->lastAllocPageSize);
@@ -166,7 +166,7 @@ mempage* region::newpage(mempage* succ, unsigned int sz) {
   return p;
 }
 
-void region::allocpage(unsigned int sz) {
+void region::allocpage(size_t sz) {
   if (this->freep != 0 && sz <= this->freep->size) {
     // used = head free : used
     mempage* usednp = this->freep;
